feat: validaciones de payload ampliado (15 campos) en DatosRegistroAnalisis, ajuste de tipo en Entity

+++ backend/backend-api/src/main/java/com/hackathon/energia/model/AnalisisEnergetico.java
@@ -44,7 +44,7 @@ public class AnalisisEnergetico {
     private String tipoInmueble;
 
     @Column(name = "horas_alto_consumo")
-    private Integer horasAltoConsumo;
+    private Double horasAltoConsumo;
 
     @Column(name = "categoria")
     @Enumerated(EnumType.STRING)
